feat: example - GET default.config
