+++ default.conf
@@ -3,7 +3,7 @@ http {
 	client_max_body_size 50m;
 	server {
 		root /www;
-		server_name test;
+		server_name localhost;
 		client_max_body_size 50m;
 		listen 8080;
 		location / {
