+++ src/ClientEventControllerRead.cpp
@@ -110,6 +110,112 @@ void ClientEventController::parseBody() {
   }
 }
 
+static ServerConfig *selectServerConfig(
+    RequestVO request, std::vector<ServerConfig *> serverConfigs) {
+  if (request.hasHeader("Host") == false) {
+    return NULL;
+  }
+  std::string host = request.getHeader("Host");
+  host = host.substr(0, host.find(":"));
+  if (serverConfigs.size() == 0) {
+    return NULL;
+  }
+  std::vector<ServerConfig *> candidates;
+  for (size_t i = 0; i < serverConfigs.size(); i++) {
+    if (serverConfigs[i]->getServerName() == host) {
+      candidates.push_back(serverConfigs[i]);
+    }
+  }
+  ServerConfig *config = serverConfigs[0];
+  if (candidates.size() > 0) {
+    config = candidates[0];
+  }
+  return config;
+}
+
+static std::vector<std::string> strSplit(const std::string &str, char delim) {
+  std::vector<std::string> result;
+  size_t prev = 0;
+  for (size_t i = 0; i < str.size(); i++) {
+    if (str[i] == delim) {
+      result.push_back(str.substr(prev, i - prev));
+      prev = i + 1;
+    }
+  }
+  if (prev < str.size()) {
+    result.push_back(str.substr(prev, str.size() - prev));
+  }
+  return result;
+}
+
+static bool isParentPath(const std::string &parent, const std::string &child) {
+  if (parent.size() > child.size()) {
+    return false;
+  }
+  const std::vector<std::string> parentSplit = strSplit(parent, '/');
+  const std::vector<std::string> childSplit = strSplit(child, '/');
+  if (parentSplit.size() > childSplit.size()) {
+    return false;
+  }
+  for (size_t i = 0; i < parentSplit.size(); i++) {
+    if (parentSplit[i] != childSplit[i]) {
+      return false;
+    }
+  }
+  return true;
+}
+
+static const LocationConfig *selectLocationConfig(
+    const std::vector<LocationConfig> &locations, const std::string &uri) {
+  const LocationConfig *res = NULL;
+  size_t maxLocationLen = 0;
+  for (size_t i = 0; i < locations.size(); i++) {
+    const std::string &locationURI = locations[i].getUri();
+    if (isParentPath(locationURI, uri) == false) {
+      continue;
+    }
+    if (maxLocationLen > locationURI.size()) {
+      continue;
+    }
+    if (locationURI[locationURI.size() - 1] != '/' && locationURI != uri) {
+      continue;
+    }
+    maxLocationLen = locationURI.size();
+
+    res = &locations[i];
+  }
+  return res;
+}
+
+void ClientEventController::beginProcess(int statusCode) {
+  evSet(EVFILT_READ, EV_DELETE);
+  statusCode_ = statusCode; 
+  if (statusCode >= 200) {
+    // TODO: 적절한 processor 처리 필요.
+    std::cout << "error code: " << statusCode << std::endl;
+    return;
+  }
+  if (config_ == NULL) {
+    ServerConfig *serverConfig =
+        selectServerConfig(request_, getServerConfigs());
+    if (serverConfig == NULL) {
+      statusCode_ = 400;
+      // TODO: 적절한 processor 처리 필요.
+      return;
+    }
+    config_ = selectLocationConfig(serverConfig->getLocationConfigs(),
+                                   strTrim(request_.getUri()));
+    if (config_ == NULL) {
+      statusCode_ = 404;
+      // TODO: 적절한 processor 처리 필요.
+      return;
+    }
+    std::cout << "debug - selected location path: " << config_->getUri() << std::endl;
+  }
+  processor_ = RequestProcessorFactory::createRequestProcessor(
+      request_, config_, kq_, this);
+}
+
 /* startline, header : 제대로 들어왔는지, 클라이언트가 무엇을 원하는지 확인. */
 /* body는 필요한 순간에 읽는다, 여기선 header 내용만 검사한다. */
 /* 하나의 요청이 여러 개로 찢어져서 들어온 건지, 다 들어왔는지도 확인. */
@@ -143,9 +249,7 @@ enum EventController::returnType ClientEventController::clientRead(
         parseHeader();
       } catch (std::exception &e) {
         std::cout << "Error: " << e.what() << std::endl;  // debug
-        statusCode_ = 401;
-        evSet(EVFILT_READ, EV_DELETE);
-        evSet(EVFILT_WRITE, EV_ADD);
+        beginProcess(401);
         return PENDING;
       }
       std::map<std::string, std::string>::const_iterator it =
@@ -155,14 +259,14 @@ enum EventController::returnType ClientEventController::clientRead(
         double contentLen = std::strtod(it->second.c_str(), &end);
         if ((end && *end != '\0') || contentLen < 0) {
           statusCode_ = 402;
+          // beginProcess(402); throw 가 필요한가?
           throw std::invalid_argument("wrong Content-Length format");
         }
         contentLength_ = static_cast<size_t>(contentLen);
       }
       parseBody();
       printParseResult();
-      evSet(EVFILT_READ, EV_DELETE);
-      evSet(EVFILT_WRITE, EV_ADD);
+      beginProcess(0);
     }
     return PENDING;
   } else {
@@ -171,14 +275,11 @@ enum EventController::returnType ClientEventController::clientRead(
       parseBody();
     } catch (std::exception &e) {
       std::cout << "Error: " << e.what() << std::endl;  // debug
-      statusCode_ = 401;
-      evSet(EVFILT_READ, EV_DELETE);
-      evSet(EVFILT_WRITE, EV_ADD);
+      beginProcess(401);
       return PENDING;
     }
   }
   printParseResult();
-  evSet(EVFILT_READ, EV_DELETE);
-  evSet(EVFILT_WRITE, EV_ADD);
+  beginProcess(0);
   return PENDING;
 }
